+++ src/components/ui/textinput.tsx
@@ -1,28 +1,27 @@
-// src/components/Text.tsx
 import { useRef, useState } from 'react';
 
-type TextStatus = 'inactive' | 'typing' | 'error' | 'active';
+type TextinputStatus = 'inactive' | 'typing' | 'error' | 'active';
 
-interface TextProps {
+interface TextinputProps {
   value?: string;
   onChange?: (value: string) => void;
   placeholder?: string;
-  status?: TextStatus;
+  status?: TextinputStatus;
   className?: string;
   maxLength?: number;
 }
 
-const Text = ({
+const Textinput = ({
   value = '',
   onChange,
-  placeholder = 'Text',
+  placeholder = 'Textinput',
   status,
   className = '',
   maxLength = 15, 
-}: TextProps) => {
+}: TextinputProps) => {
   const textRef = useRef<HTMLInputElement>(null);
   const [hasBlurred, setHasBlurred] = useState(false);
-  const [currentStatus, setCurrentStatus] = useState<TextStatus>('inactive'); // 상태 추적
+  const [currentStatus, setCurrentStatus] = useState<TextinputStatus>('inactive'); // 상태 추적
 
   const handleBlur = () => {
     setHasBlurred(true);
@@ -33,7 +32,7 @@ const Text = ({
   };
 
   // 상태 감지
-  const derivedStatus: TextStatus = (() => {
+  const derivedStatus: TextinputStatus = (() => {
     if (status) return status;
     if (currentStatus === 'error') return 'error'; // error 상태 유지
     if (value.length === 0) return 'inactive'; // 초기 상태
@@ -93,4 +92,4 @@ const Text = ({
   );
 };
 
-export default Text;
+export default Textinput;
