+++ src/pages/onboarding/Nickname.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useNavigate } from 'react-router';
 import ProgressHeader from '@/components/ui/progressheader';
 import { Button } from '@/components/ui/button';
-import Input from '@/componenets/Input/Input';
+import Input from '@/components/ui/input';
 
 export const Nickname: React.FC = () => {
   const [nickname, setNickname] = useState('');
